feat(front-end): ask users to confirm delete
When clicking on the delete (trash) icon or button, there was no
warning or double-checking before deleting.  This can easily cause a
section, education, experience, etc. to be deleted.  To remedy this,
after clicking on a trash icon once, it turns red.  Then after clicking
on the icon again while it is red, will the delete happen.

A custom hook called useConfirmClick is created to handle the logic for
confirming a click before running a function.

TrashIcon and TrashButton have been created to abstract the original
SVG icon or button code into simple single components.  They contain
the code to display the icon or button, as well as utilize
useConfirmClick to run the logic.

For TrashButton, the CSS styling had to be directly placed in the JSX
return, which directly applies the CSS inline to HTML elements, because
the Bootstrap overrides using "!important" and Bootstrap CSS were not
allowing the styling in associated CSS files to have precedence.

Note that the padding for TrashIcon and TrashButton had to be adjusted
when they turn red, because a question mark character gets added, else
clicking once will change the size of the trash SVG or the button size.

The trash.svg file is converted to a React component called TrashSvg,
so that colors can be applied to it.  Consequently, trash.svg has been
deleted.

All uses of the trash SVG have been replace with TrashIcon and
TrashButton.

Written tests have been skipped due to time, but basic manual tests
have been performed.