refactor(service): consolidate create function
In educationService.js and experienceService.js, createEducation and
createExperience were nearly identical, because the logic for any
section type should be doing the same thing.

The code and unit tests for creating a section item has now been
consolidated and generalized into commonSectionsService files.

For the unit tests, note that Objects, representing a section type and
its document relationship, can be used instead of creating Classes or
any fancy Jest mocks, because Classes are essentially Objects.