refactor(service): consolidate create relationship function
In the education and experience services, consolidated identical code
related to creating document-education and document-experience
relationships into the createDocumentXSectionTypeRelationship function
in commonSectionsService.js.

Unit tests were rewritten, because one of them had test cases that were
irrelevant to the logic, and there was a missing test for when the
"add" method of the document-(section type) relationship class throws
an unknown error.