chore(front-end): only change contact info in master
The contact information was able to be edited in non-master resumes,
which was inconsistent, because nothing else in the those resumes were
able to be edited, added, or permanently be deleted.

The pencil/edit icon for the other resumes has now been removed and
only shows up in the master resume.  The CSS had to also be changed to
accommodate for an empty header.

Written tests have been skipped due to time, but basic manual tests
have been performed.