chore: remove updateAllPositions duplicate code
In resume-manager-backend\models\relationship.js, the
updateAllPositions method had code that checks if the number of items
being repositioned is equal to all of those kinds of items in a
particular document or experience.  This check is already done or
should be done in the service layer, which also includes or should
include a check for if all of the exact items are present in the list
to be repositioned.  Functions in models should only be concerned about
creating, getting, updating, or deleting data in the database, not
whether the data is valid.

Updated tests accordingly.  So far, only Document_X_Section has a test
for that piece of code.