refactor(back-end): use deleteItem in services
Refactored delete code for document, education, experience, and text
snippet services and replaced it with deleteItem from commonServices.js.

For the deleteDocument function in the document service, all but one of
the original unit tests have been removed, because the tests are now
performed in commonServices.test.js.  The remaining tests are for the
logic in deleteDocument.