test(front-end): add useNotifications tests
Note that in the "removes an alert." test, when calling the addAlert
function in the for-loop, the code execution flow jumps to the next
test, because there is a synchronicity requirement even though the code
in useNotifications does not show it.  This also occurs in the next
test, "clears all alerts.".  This has resulted in requiring all of the
alert functions to be awaited, so that each test is completed before
moving to another one.